// src/app/Model/Customer.model.ts
// customer.model.ts
export interface Customer {
    loginTime: any;
    logoutTime:any;
    customerName: string;
    aadhaarNumber: string;
    accountNumber:string;
    accountBalance:string;
    mobileNumber: string;
    dateOfBirth: string; 
    userName: string;
    password: string;
    //confirmPassword: string;
  }
  